requirements.txt
requests>=2.32.0
pandas>=2.2.0
matplotlib>=3.10.0
seaborn>=0.13.0
numpy>=2.0.0
boto3>=1.38.0
python-docx>=1.1.0
openai>=1.0.0
python-dotenv>=1.0.0
jupyter>=1.0.0
notebook>=6.5.0
openpyxl>=3.1.0
